:art: Huge refactor 2
