+++ chess/core/query.py
