+++ tests/special_moves/en_passant_test.py
@@ -1,5 +1,5 @@
 from chess.core.moving import move
-from chess.core.destinations import destinations
+from chess.core.query import destinations
 from chess.core.models import Piece, Coordinate
 from chess.core.utils import empty_board, set_at, piece_at
 from chess.core.game import Game
